Drop callback args in favor of single, uniform test convention
This one's largely a cosmetic change. We remove callback args/worker in
favor of using test case local args using `JobArgsReflectKind` and a
standard invocation to `AddWorker`:

    type JobArgs struct {
        JobArgsReflectKind[JobArgs]
    }

    AddWorker(config.Workers, WorkFunc(func(ctx context.Context, job *Job[JobArgs]) error {
        return JobCancel(errors.New("oops"))
    }))

It's a little more verbose, but has the advantage of using only normal
River primitives, meaning that one has to be familiar with fewer test
helpers to understand what's going on.

Besides that and convention, another small benefit is that it introduces
a little more type safety and makes job args impossible to accidentally
cross reference between test cases. Previously, a test could insert a
callback args while intending to use a callback for them registered in a
different test case. With the new approach, it's very obvious when
`JobArgs` is either forgotten (a compilation error), or when there's a
missing `AddWorker` (because that's supposed to live close to the
definition of `JobArgs`).