Require standard format for job kinds, mainly to avoid commas and spaces
Here, add a client configuration check that requires job kinds adhere to
a standard format, mainly in an attempt to avoid commas and spaces to
make inputs more predictable for an upcoming search UI.

For compatibility purposes, for the time being the check can be disabled
with the new client option `SkipJobKindValidation`, although this option
will be removed in the future.

The character set I've allowed is a little arbitrary (although a lot of
the characters are there to let `JobArgsReflectKind` keep working
properly), so we could extend it to more characters before release if we
get feedback on it.