Driver tests for full `lib/pq` connection + necessary fixes
So it turns out that while bearing some relation to `lib/pq` through
various pieces that sqlc pulls in, the `riverdatabasesql` driver didn't
actually work when used with `lib/pq`. When testing it, we'd open a
`sql.DB` pool _via_ Pgx, so Pgx was doing all the heavy lifting.

Here, add an alternate driver test strategy to test `riverdatabasesql`
with a `lib/pq` pool, thereby ensuring that the full range of operations
works even with no Pgx in the mix at all.

Things were broken, so we get those fixed up. There we two main
problems:

* As described in #545, `lib/pq` doesn't handle intervals correctly, and
  reads them in as ints, which is a big problem because they're in
  nanoseconds. We work around this by injecting intervals (like leader
  TTLs) an second floats instead, which really doesn't feel that much
  worse.

* The custom bits class for `unique_states` didn't work. To fix this I
  move to an alternate strategy of inserts `unique_states` as an
  integer, a technique that I've been using in SQLite because there's no
  `bits` type. I could make a strong argument that the use of an integer
  over `bits` would've been a better design from the get go anyway,
  because it makes all the bit-wise arithmetic much simpler as you use
  C-style conventions like `states & (1 << 3)` that are very widespread
  and well understood. Either way though, things work reasonably well if
  we insert as an integer, then transition to storage in bits.

Fixes #545 and #566. Supersedes #882.