Allow args to driver's raw `Exec`
Currently, unlike underlying connections/transactions, the driver
version of `Exec` doesn't take parameters. This is mostly okay because
`Exec` is mainly needed for migrations, but as I was trying to convert
some code over in Pro's test suite, I did find that it'd be kind of nice
to be able to use parameters there too so we can use an executor instead
of raw transaction.

This is probably okay because the executor version of `QueryRow` already
lets parameters be passed in. There is a possible problem wherein
different databases use different placeholders (Postgres's `$` versus
SQLite's `?` for example), but when we solve that we'll have to solve it
for `QueryRow` anyway.