Suppress producer error log on shutdown
I noticed when running some benchmarking tonight that I'd occasionally
get an error line at the end even though things seem to finish cleanly:

    May 13 18:13:36.622 ERR producer: Error fetching queue settings err="context canceled"

Looking at the implementation, it looks like this is something that
could happen randomly in the poll-only path. If the producer happens to
be polling for queue changes as shutdown happens, it'll log an error as
the context is cancelled. This is more likely to happen for SQLite
because we constrain the connection pool to only one active connection.

Here, only produce an error log conditionally in cases where the parent
context has not been cancelled.