Use simply `old_name` in job rename example
In the docs for `JobArgsWithKindAliases` we walk through an example of
renaming `old_name` to `new_name`, but in the phase where aliases are
added, we have the old name `old_name_still_recognized`.

As I writing up docs for renaming jobs, this didn't really feel like it
made much sense to me. We're supposed to be showing that `old_name` gets
moved into `KindAliases`, but if `old_name` changes to something else,
it defeats the clarity of the example such that it doesn't make all that
much sense anymore.

Here, use `old_name` in `KindAliases` instead.