Truncate tables across all migration lines in `riverdbtest`
Fixes a bug in `riverdbtest` where in the presence of multiple migration
lines, we'd overwrite our set of truncation tables on every iteration of
the loop, so we'd only end up keeping the last one. Here, make sure to
append all tables so they're all included.

This bug doesn't manifest in the main project because we never have more
than one migration line, but we do see it causing trouble in the Pro
test suite after fixing a bug there in `GetMigrationTruncateTables`.

I didn't add a test for this even though it might be useful because we
unfortunately just don't have any good way of doing it at the moment.
We'd need to have some kind of mock driver to pull this off.