Add `--schema` option to `river migrate-get` + hide schema comments otherwise
Here, add a `--schema` option to the `river migrate-get` command that
lets the caller inject a custom schema into the SQL generated by the
command. If the `--schema` command is not provided, we hide the `/*
TEMPLATE: schema */` comments for cleanliness (previously, they were
returned, which wasn't harmful but added visual noise).

Since we're making changes anyway, add a test suite for `migrate-get`.
It should've been there before anyway and it will help protect against
regressions.

Only tangentially related, fix a small bug in the migration "flags" that
are output like `-- River main migration 001 [up]` in that when using
the default "main" line, the line was being accidentally omitted.

Also unrelated, but add missing docs for `river migrate-list` which have
apparently accidentally been a TODO for quite some time.

Fixes #902.