Correct reindexer's handling of explicit schema
Here, correct the reindexer's handling of an explicitly inject schema,
which was being added with a "." _before_ the schema name instead of
after.

We also patch up the test suite. Looking into this, the error was being
emitted in the tests as clear as day, but none of the tests were
asserting anything (basically just that the reindexer ran in some way,
shape, or form, not that it had been successful or done any work), so it
was just being ignored.

Fixes #915.