Push `REINDEX` down into driver so reindexer works on SQLite (#911)
I realized when #909 that the reindexer isn't currently compatible with
the SQLite driver because SQLite's `REINDEX` syntax is a little
different, and it doesn't support the `CONCURRENTLY` keyword. The
reindexer runs once an hour so nothing caught the problem.

Here, push the reindex command down into the driver and add a test for
it. This is a good time to do this since we're changing the driver
interface by the necessity for the next version anyway so we're not
breaking anything that's not already broken.

One outstanding issue is that SQLite can't reindex concurrently so the
reindexer would fully lock the database while it's reindexing the
configured indexes. There's nothing that can be done about this for now,
so we might have to recommend that users running a larger database
disable the reindexer if they're using SQLite.