Return error about job list metadata on SQLite on `JobListTx` (#924)
Fix a problem reported in #923. We return a special error on SQLite when
someone tries to use `JobListParams.Metadata` because we can't support
the Postgres `@>` operator, which doesn't exist in any other database.
Somehow, this error was only returned on `JobList` and not `JobListTx`.

Here, bring it to `JobListTx` too. Add a new set of client driver tests
for `JobListTx` so we're doing reasonable exercise the function under
SQLite.