Out of transaction notification on job insert and queue changes
Here, we try to implement a system that'll make drivers that don't
support listen/notify like SQLite more responsive by default (i.e.
without having to crank their fetch intervals down to miniscule numbers
and have them fetch with hyper-frequency) so they respond to a new job
or queue being paused quickly.

Currently, drivers with listeners send out notifications which are then
received by the producer so that it can enact work. Drivers without
listeners are left out in the cold though because they don't have any
good listen/notify alternative.

This proposal takes non-tx functions like `Client.Insert` or
`Client.QueuePause` and makes a special call to the producer to notify
it of what happened so it can respond quickly without waiting for a
poll.

The approach of course has a couple limitations:

* It only works for non-tx functions. If we were to notify the producer
  within a transaction, the producer would wake up and find nothing to
  do because the relevant data hasn't been persisted yet. This is taken
  care of on the listener drivers because listen/notify is transaction
  friendly.

* Only works within the same process. A client obviously can't notify a
  producer in a different Go process, so those will still have to poll.
  Still, if even one producer can get a notification about an inserted
  job immediately, that job still gets worked faster.

    This improvement is also aimed pretty specifically aimed at our test
    suite, where there's only ever one process running. Being able to
    notify a producer within the same process means that we don't need
    to put an aggressive poll loop into ever example or client test that
    uses SQLite.

Longer term, I'd like to replace this with something else, like a
general purpose unlogged table that could process notifications in a
transaction in a similar way to how listen/notify does it (although I'd
have to do a little more work on this, I'm not sure it'd work), but I
figure that this might be good enough for the time being. This doesn't
change the public API at all, so it should be fairly easy to replace all
changes made here with something more suitable if it comes up.