Reindexer: Skip work if artifacts from a failed reindex are present + increase timeout
Here, we attempt to resolve #909, in which a user had found a great
number of invalid indexes in their database, caused by the reindex
running, having its context cancelled on timeout, and leaving an invalid
index behind. 24 hours later it would try again, with the same result.

We add a check to the reindexer to look for artifacts of previously
failed reindex operations (or concurrently running reindex operations,
though that shouldn't happen due to the leadership system), and if
found, skip reindexing in favor of logging a warning about what
happened.

As documented by Postgres [1], a concurrent reindex creates a new index
with a prefix of `*_ccnew`, and swaps it into place if reindexing
succeeds. The existing index is renamed to `*_ccold` and dropped
concurrently. We look for both these suffixes, and skip reindexing if we
find either of them.

I opted not to clean reindexing artifacts for the same reason that we're
skipping reindexing if finding them in the first place: even after
cleaning them out, reindexing is likely to fail again. If it failed once
at a given timeout threshold, then it's probably fated to fail again.

When running and failing multiple times, reindexing will add numbers to
each new attempt like `*_ccnew1`, `*_ccnew2`, ... `*_ccnew10`,
`*_ccnew11`. I opted not to look for these because if any of them
exists, we can generally assume that the first `*_ccnew` does too. If
not (say a user manually dropped it), the reindexing will occur, fail
again, and `*_ccnew` will be there. I don't think we have to worry about
this edge too much.

Lastly, I also increase the reindex context timeout from 15 seconds to 1
minute. As seen in #909, a large table will have reindex builds that
take at least 45 seconds, so 1 minute is that plus some margin. Use of
`CONCURRENTLY` should ensure safe continued table access even while a
longer reindex is running. Even 1 minute might not be long enough for
every case, but it's a lot, and should reduce the number of users who
see failures dramatically (from few users to very, very few users). The
choice of 15 seconds originally was quite arbitrary and there's no
reason to be married to it.

Fixes #909.

[1] https://www.postgresql.org/docs/current/sql-reindex.html#SQL-REINDEX-CONCURRENTLY