Better error in panic when accessing periodic jobs bundle in client that won't work (#938)
Related to #937. If an insert-only client is created and the
`PeriodicJobs()` bundle is accessed, it'll panic on a nil pointer error.
The error is appropriate because if the client will never work jobs, it
can never be started, which means it can never win leader election,
which means that adding periodic jobs to it will never have any effect,
but the bad error message is definitely not ideal.

Here, detect this condition and panic with an error that better explains
the situation to the caller.

Fixes #937.