tweak config metadata updates, fix adding existing queue (#929)
* tweak config metadata updates, fix adding existing queue

Some small changes to the pilot integration APIs to facilitate
equivalent Pro changes. While in here, I realized it's possible to add a
queue that already exists, an undefined behavior which could result in a
leaked producer. To fix that, I introduced an exported error type and
made use of it on this API.

* fix duplicate drivers in tests