Extract `mapSliceError` into `rivershared` `sliceutil.MapError` function
I just noticed that we end up redefining `mapSliceError` in every
driver. Here, extract it into `rivershared` as `sliceutil.MapError` so
it can be reused everywhere. It's a good time to do this anyway because
we're about to have a mandatory update across project boundaries so Pro
will be able to get access to it too.